server/pyproject.toml
[project]
name = "metakb"
authors = [
    {name = "Alex H Wagner"},
    {name = "Kori Kuzma"},
    {name = "James Stevenson"},
    {name = "Brian Walsh"},
    {name = "Jeff Liu"},
    {name = "Matthew Cannon"},
    {name = "Katie Stahl"},
]
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.11"
description = "A search interface for cancer variant interpretations assembled by aggregating and harmonizing across multiple cancer variant interpretation knowledgebases."
license = {file = "LICENSE"}
dependencies = [
    "ga4gh.vrs>=2.1.3,<3.0",
    "ga4gh.cat_vrs~=0.6.0",
    "ga4gh.va_spec~=0.3.0",
    "gene-normalizer[etl]~=0.10.2",
    "variation-normalizer~=0.14.1",
    "disease-normalizer[etl]~=0.11.2",
    "thera-py[etl]~=0.11.0",
    "civicpy~=3.1",
    "requests",
    "pydantic==2.*",
    "pydantic-settings",
    "requests-cache",
    "neo4j==5.*",
    "uvicorn",
    "fastapi",
    "boto3",
    "botocore",
    "asyncclick",
]
dynamic = ["version"]

[project.optional-dependencies]
tests = [
    "pytest",
    "pytest-cov",
    "mock",
    "pytest-asyncio",
    "deepdiff",
    "httpx",
    "jsonschema~=4.24",  # pin to avoid deprecation of ref resolver
    "pyyaml",
]
dev = [
    "pre-commit>=4.2.0",
    "ruff==0.12.1",
    "fastapi[standard]",
]
notebooks = ["ipykernel", "jupyterlab"]
docs = [
    "sphinx==6.1.3",
    "sphinx-autodoc-typehints==1.22.0",
    "sphinx-autobuild==2021.3.14",
    "sphinx-copybutton==0.5.2",
    "sphinxext-opengraph==0.8.2",
    "furo==2023.3.27",
    "sphinx-github-changelog==1.2.1",
    "sphinx-click==5.0.1",
    "sphinxcontrib-images==0.9.4",
]

[project.urls]
Homepage = "https://github.com/cancervariants/metakb"
Documentation = "https://github.com/cancervariants/metakb"
Changelog = "https://github.com/cancervariants/metakb/releases"
Source = "https://github.com/cancervariants/metakb"
"Bug Tracker" = "https://github.com/cancervariants/metakb/issues"

[project.scripts]
metakb = "metakb.cli:cli"

[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
root = ".."

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
metakb = ["*.cypher"]

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report term-missing --ignore=codebuild/"
testpaths = ["tests"]
markers = [
    "ci: tests safe to run in CI",
]

[tool.coverage.run]
branch = true

[tool.ruff]
src = ["src"]
exclude = ["../docs/*", "codebuild/*"]

[tool.ruff.lint]
select = [
    "F",  # https://docs.astral.sh/ruff/rules/#pyflakes-f
    "E", "W",  # https://docs.astral.sh/ruff/rules/#pycodestyle-e-w
    "I",  # https://docs.astral.sh/ruff/rules/#isort-i
    "N",  # https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "D",  # https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "UP",  # https://docs.astral.sh/ruff/rules/#pyupgrade-up
    "ANN",  # https://docs.astral.sh/ruff/rules/#flake8-annotations-ann
    "ASYNC",  # https://docs.astral.sh/ruff/rules/#flake8-async-async
    "S",  # https://docs.astral.sh/ruff/rules/#flake8-bandit-s
    "B",  # https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "A",  # https://docs.astral.sh/ruff/rules/#flake8-builtins-a
    "C4",  # https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "DTZ",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
    "T10",  # https://docs.astral.sh/ruff/rules/#flake8-datetimez-dtz
    "EM",  # https://docs.astral.sh/ruff/rules/#flake8-errmsg-em
    "LOG",  # https://docs.astral.sh/ruff/rules/#flake8-logging-log
    "G",  # https://docs.astral.sh/ruff/rules/#flake8-logging-format-g
    "INP",  # https://docs.astral.sh/ruff/rules/#flake8-no-pep420-inp
    "PIE",  # https://docs.astral.sh/ruff/rules/#flake8-pie-pie
    "T20",  # https://docs.astral.sh/ruff/rules/#flake8-print-t20
    "PT",  # https://docs.astral.sh/ruff/rules/#flake8-pytest-style-pt
    "Q",  # https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "RSE",  # https://docs.astral.sh/ruff/rules/#flake8-raise-rse
    "RET",  # https://docs.astral.sh/ruff/rules/#flake8-return-ret
    "SLF",  # https://docs.astral.sh/ruff/rules/#flake8-self-slf
    "SLOT",  # https://docs.astral.sh/ruff/rules/#flake8-slots-slot
    "SIM",  # https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "ARG",  # https://docs.astral.sh/ruff/rules/#flake8-unused-arguments-arg
    "PTH",  # https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
    "PGH",  # https://docs.astral.sh/ruff/rules/#pygrep-hooks-pgh
    "FAST",  # https://docs.astral.sh/ruff/rules/#fastapi-fast
    "PLC",  # https://docs.astral.sh/ruff/rules/#convention-c
    "PLE",  # https://docs.astral.sh/ruff/rules/#error-e_1
    "TRY",  # https://docs.astral.sh/ruff/rules/#tryceratops-try
    "PERF",  # https://docs.astral.sh/ruff/rules/#perflint-perf
    "FURB",  # https://docs.astral.sh/ruff/rules/#refurb-furb
    "RUF",  # https://docs.astral.sh/ruff/rules/#ruff-specific-rules-ruf
]
fixable = [
    "I",
    "F401",
    "D",
    "UP",
    "ANN",
    "B",
    "C4",
    "LOG",
    "G",
    "PIE",
    "PT",
    "RSE",
    "SIM",
    "FAST",
    "PLC",
    "PLE",
    "TRY",
    "PERF",
    "FURB",
    "RUF"
]
# ANN002 - missing-type-args
# ANN003 - missing-type-kwargs
# D203 - one-blank-line-before-class
# D205 - blank-line-after-summary
# D206 - indent-with-spaces*
# D213 - multi-line-summary-second-line
# D300 - triple-single-quotes*
# D400 - ends-in-period
# D415 - ends-in-punctuation
# E111 - indentation-with-invalid-multiple*
# E114 - indentation-with-invalid-multiple-comment*
# E117 - over-indented*
# E501 - line-too-long*
# W191 - tab-indentation*
# S321 - suspicious-ftp-lib-usage
# PLC0206 - dict-index-missing-items
# *ignored for compatibility with formatter
ignore = [
    "ANN002", "ANN003",
    "D203", "D205", "D206", "D213", "D300", "D400", "D415",
    "E111", "E114", "E117", "E501",
    "W191",
    "S321",
    "PLC0206",
]

[tool.ruff.lint.per-file-ignores]
# ANN001 - missing-type-function-argument
# ANN2 - missing-return-type
# S101 - assert
# B011 - assert-false
# D100 - undocumented-public-module
# D102 - undocumented-public-class
# D103 - undocumented-public-function
# D104 - undocumented-public-package
# INP001 - implicit-namespace-package
# ARG001 - unused-function-argument
# SLF001 - private-member-acces
# N815 - mixed-case-variable-in-class-scope
"tests/*" = [
    "ANN001",
    "ANN2",
    "S101",
    "B011",
    "D100",
    "D102",
    "D103",
    "D104",
    "INP001",
    "SLF001",
    "ARG001"
]
"src/metakb/schemas/*" = ["ANN102", "N815"]
"src/metakb/repository/queries/catalog.py" = ["D102"]

[tool.ruff.lint.flake8-annotations]
mypy-init-return = true

[tool.ruff.format]
docstring-code-format = true
